refactor(swarm_team): use simpler generic models
